+++ WinApiLecture/Client/CBtnUI.cpp
@@ -55,7 +55,7 @@ void CBtnUI::render(HDC _dc)
 	CUI::render(_dc);
 
 	//UI버튼의 사각형 공간 정보 저장(finalpos로 저장된 좌표가 상대좌표이기에, 상속관계의 버튼의 정상적 위치 받아올 수 있음)
-	RECT rt = { GetFinalPos().x, GetFinalPos().y, GetFinalPos().x + GetScale().x, GetFinalPos().y + GetScale().y };
+	RECT rt = { (LONG)GetFinalPos().x, (LONG)GetFinalPos().y, (LONG)(GetFinalPos().x + GetScale().x), (LONG)(GetFinalPos().y + GetScale().y) };
 
 	//한줄 출력 + 수직 가운데 위치 + 수평 가운데에서 버튼 UI의 텍스트 출력
 	DrawText(_dc, m_strBtnText.c_str(), -1, &rt, DT_SINGLELINE | DT_VCENTER | DT_CENTER);
